+++ library/src/main/java/com/prolificinteractive/materialcalendarview/CalendarPagerView.java
@@ -24,6 +24,8 @@ import static java.util.Calendar.DAY_OF_WEEK;
 abstract class CalendarPagerView extends ViewGroup implements View.OnClickListener {
 
     protected static final int DEFAULT_DAYS_IN_WEEK = 7;
+    protected static final int DEFAULT_MAX_WEEKS = 6;
+    protected static final int DAY_NAMES_ROW = 1;
     private static final Calendar tempWorkingCalendar = CalendarUtils.getInstance();
     private final ArrayList<WeekDayView> weekDayViews = new ArrayList<>();
     private final ArrayList<DecoratorResult> decoratorResults = new ArrayList<>();
@@ -88,26 +90,6 @@ abstract class CalendarPagerView extends ViewGroup implements View.OnClickListen
 
     protected int getFirstDayOfWeek() {
         return firstDayOfWeek;
-    }
-
-    public void setFirstDayOfWeek(int dayOfWeek) {
-        this.firstDayOfWeek = dayOfWeek;
-
-        Calendar calendar = resetAndGetWorkingCalendar();
-        calendar.set(DAY_OF_WEEK, dayOfWeek);
-        for (WeekDayView dayView : weekDayViews) {
-            dayView.setDayOfWeek(calendar);
-            calendar.add(DATE, 1);
-        }
-
-        calendar = resetAndGetWorkingCalendar();
-        for (DayView dayView : dayViews) {
-            CalendarDay day = CalendarDay.from(calendar);
-            dayView.setDay(day);
-            calendar.add(DATE, 1);
-        }
-
-        updateUi();
     }
 
     protected abstract void buildDayViews(Collection<DayView> dayViews, Calendar calendar);
@@ -207,8 +189,8 @@ abstract class CalendarPagerView extends ViewGroup implements View.OnClickListen
     @Override
     public void onClick(View v) {
         if (v instanceof DayView) {
-            DayView dayView = (DayView) v;
-            mcv.onDateClicked(dayView.getDate(), !dayView.isChecked());
+            final DayView dayView = (DayView) v;
+            mcv.onDateClicked(dayView);
         }
     }
 
@@ -240,28 +222,36 @@ abstract class CalendarPagerView extends ViewGroup implements View.OnClickListen
         }
 
         //The spec width should be a correct multiple
-        final int measureTileSize = specWidthSize / DEFAULT_DAYS_IN_WEEK;
+        final int measureTileWidth = specWidthSize / DEFAULT_DAYS_IN_WEEK;
+        final int measureTileHeight = specHeightSize / getRows();
 
         //Just use the spec sizes
         setMeasuredDimension(specWidthSize, specHeightSize);
 
         int count = getChildCount();
+
         for (int i = 0; i < count; i++) {
             final View child = getChildAt(i);
 
             int childWidthMeasureSpec = MeasureSpec.makeMeasureSpec(
-                    measureTileSize,
+                    measureTileWidth,
                     MeasureSpec.EXACTLY
             );
 
             int childHeightMeasureSpec = MeasureSpec.makeMeasureSpec(
-                    measureTileSize,
+                    measureTileHeight,
                     MeasureSpec.EXACTLY
             );
 
             child.measure(childWidthMeasureSpec, childHeightMeasureSpec);
         }
     }
+
+    /**
+     * Return the number of rows to display per page
+     * @return
+     */
+    protected abstract int getRows();
 
     /**
      * {@inheritDoc}