+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 Change Log
 ==========
+Version 1.3.0 *(2016-05-16)*
+----------------------------
+
+* New: MCV `goToNext` and `goToPrevious` API to programmatically trigger paging
+* New: Allow users to click on dates outside of current month with `setAllowClickDaysOutsideCurrentMonth`
+* New: Set tile width/height separately rather than single tile size
+* New: Attributes: mcv_tileWidth, mcv_tileHeight, mcv_calendarMode
+* Change: `CalendarMode.WEEK` officially marked `@Experimental`, use with caution
+* Change: `getTileSize` is deprecated, use `getTileWidth` and `getTileHeight`. `setTileSize` still works as a convenience method to set width and height at the same time.
+* Fix: Issue with arrow not enabled when setting maxDate
+* Fix: Issue with number of pages not calculated correctly with maxDate causing last page to be unreachable
+* Fix: TalkBack content descriptions for pager view, forward/back arrows, and ability to set them manually
+* Fix: Crash while in Week mode when `CalendarPagerAdapter#getItemPosition` is called
+* Fix: Calendar Mode is retained on restore instance state
+* Fix: Min/Max date range is retained on restore instance state
+* Issue: Week mode - Restore instance state shows the previous week of the one that was saved
+* Issue: Week mode - Some combinations of first day of week, min/max date can cause the last week not to be pagable
 
 Version 1.2.1 *(2016-05-05)*
 ----------------------------
@@ -21,7 +38,7 @@ Version 1.1.0 *(2015-10-19)*
 * New: Ability to disable month swiping with `setPagingEnabled()`
 * Fix [#149](https://github.com/prolificinteractive/material-calendarview/issues/149):
   save selected dates as a typed List instead of an array.
-* Change: Some preformance optimizations
+* Change: Some performance optimizations
 
 Version 1.0.1 *(2015-09-30)*
 ----------------------------